flask_app/requirements.txt
boto3==1.26.91
Flask==2.2.3
librosa==0.10.0
python-dotenv==1.0.0
torch==1.13.1
